% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{rmst}
\alias{rmst}
\title{Calculate the RMST in each arm and the between-group difference}
\usage{
rmst(formula, data, trunc, alpha, method = c("KM", "pseudo", "RP"), ...)
}
\arguments{
\item{formula}{A formula object, with the response on the left of a ~ operator, and the a single term for arm on the right. The response must be a survival object as returned by the Surv function.}

\item{data}{A data.frame in which to interpret the variables named in the formula.}

\item{trunc}{The time point to truncated the estimate of the RMST at.}

\item{alpha}{The default is 0.05. (1-alpha) confidence intervals are reported.}

\item{method}{The method to estimate the RMST. Note that 2 knots will be used by default when using the Royston-Parmar method unless otherwise specified.}

\item{...}{Additional arguments to be passed to the method}
}
\value{
The Restricted Mean Survival Time computated
}
\description{
This function calculates an estimate of the RMST using the specified method
}
\details{
The RMST is estimate which can be found by calculating the area under the survival curve. This function facilitates the implementation the three most discussed approaches in current literature. For further details on how the RMST is calculated for each method see the document for the appropriate method, i.e. \code{?rmstKM}, \code{?rmstPseudo}, or \code{?rmstRP}.
}
\examples{
# Load in an example data set
D <- rmst2.sample.data()

# Calculate the RMST using pseudovalues
rmst(Surv(time, status) ~ arm, data=D, trunc=5, alpha=0.05, method="pseudo")
}

