% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmstKM.R
\name{rmstKM}
\alias{rmstKM}
\title{Calculate the RMST in each arm and the between-group difference (Area under KM)}
\usage{
rmstKM(formula, data, trunc, alpha = 0.05)
}
\arguments{
\item{formula}{A formula object, with the response on the left of a ~ operator, and the a single term for arm on the right. The response must be a survival object as returned by the Surv function.}

\item{data}{A data.frame in which to interpret the variables named in the formula.}

\item{trunc}{The time point to truncated the estimate of the RMST at.}

\item{alpha}{The default is 0.05. (1-alpha) confidence intervals are reported.}
}
\value{
The Restricted Mean Survival Time computated by calculating the area under the KM curve for each arm.
}
\description{
This function calculates an estimate of the RMST by calculating the area under the KM curve in each arm. The calculations are done using a function from the \code{survRM2} package, for further information see the \code{survRM2} documentation.
}
\examples{
# Load in an example data set
D <- rmst2.sample.data()

# Calculate the RMST based on the area under the KM approach
rmstKM(Surv(time, status) ~ arm, data=D, trunc=5, alpha=0.05)
}

