% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmstPseudo.R
\name{rmstPseudo}
\alias{rmstPseudo}
\title{Calculate the RMST in each arm and the between-group difference (Pseudovalues)}
\usage{
rmstPseudo(formula, data, trunc, alpha = 0.05)
}
\arguments{
\item{formula}{a formula object, with the response on the left of a ~ operator, and the a single term for arm on the right. The response must be a survival object as returned by the Surv function.}

\item{data}{a data.frame in which to interpret the variables named in the formula.}

\item{trunc}{the time point to truncated the estimate of the RMST at.}

\item{alpha}{The default is 0.05. (1-alpha) confidence intervals are reported.}
}
\value{
The Restricted Mean Survival Time computated by fitting a GEE to pseudovalues generated for each patient
}
\description{
This function calculates an estimate of the RMST by generating a pseudovalue for each patient, and then fitting a GEE model to these pseudovalues with arm as a term in the model. The pseudovalues are generated using the function \code{pseudomean} from the package \code{pseudo} and the GEE is fit using the function \code{geese} from the package \code{geepack}. For further information see the \code{pseudo} or \code{geepack} documentation.
}
\examples{
# Load in an example data set
D <- rmst2.sample.data()

# Calculate the RMST based on the pseudovalue approach
rmstPseudo(Surv(time, status) ~ arm, data=D, trunc=5, alpha=0.05)
}

