% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmstRP.R
\name{rmstRP}
\alias{rmstRP}
\title{Calculate the RMST in each arm and the between-group difference (Royston-Parmar model)}
\usage{
rmstRP(formula, data, trunc, alpha = 0.05, knots = 2)
}
\arguments{
\item{formula}{a formula object, with the response on the left of a ~ operator, and the a single term for arm on the right. The response must be a survival object as returned by the Surv function.}

\item{data}{a data.frame in which to interpret the variables named in the formula.}

\item{trunc}{the time point to truncated the estimate of the RMST at.}

\item{alpha}{The default is 0.05. (1-alpha) confidence intervals are reported.}

\item{knots}{the number of internal knots to fit the Royston-Parmar model with.}
}
\value{
The Restricted Mean Survival Time computated by fitting Royston-Parmar model to the survival data
}
\description{
This function calculates an estimate of the RMST by fitting a Royston-Parmar model to the survival data with arm as a covariate. The model is fit using the function \code{flexsurvspline} from the package \code{flexsurv}. For further information see the \code{flexsurv} documentation. The standard errors are calculated using the Delta method.
}
\examples{
# Load in an example data set
D <- rmst2.sample.data()

# Calculate the RMST based on the Royston-Parmar model with two knots
rmstRP(Surv(time, status) ~ arm, data=D, trunc=5, alpha=0.05, knots=2)
}

