% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{helpers}
\alias{helpers}
\alias{mst_get_form_index}
\title{Helper functions for MST Assembly and Simulation}
\usage{
mst_get_form_index(x, indices, method)
}
\arguments{
\item{x}{the MST object}

\item{indices}{the input form indices}

\item{method}{the assembly method: topdown or bottomup}
}
\description{
helper functions for MST assembly and simulation

\code{mst_get_form_index} maps the input form indices to the actual indices
}
\keyword{internal}
