% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMSTdepC-package.R
\docType{package}
\name{RMSTdepC-package}
\alias{RMSTdepC}
\alias{RMSTdepC-package}
\title{RMSTdepC: Calling an Rcpp function with doParallel}
\description{
Provides an example implementation using a package to house both a C++ function and a parallelization routine.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/sjheo91/RMSTdepC}
}

}
\author{
\strong{Maintainer}: Seok-Jae Heo \email{sjheo@yuhs.ac}

}
\keyword{internal}
